move classes in respective fixed/hourly folders
