+++ scheme-service/src/main/java/uk/gov/justice/laa/fee/scheme/feecalculator/fixed/FixedFeeCalculator.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.fee.scheme.feecalculator;
+package uk.gov.justice.laa.fee.scheme.feecalculator.fixed;
 
 import uk.gov.justice.laa.fee.scheme.entity.FeeEntity;
 import uk.gov.justice.laa.fee.scheme.feecalculator.utility.FeeCalculationUtility;
