+++ scheme-service/src/main/java/uk/gov/justice/laa/fee/scheme/feecalculator/fixed/MediationFeeCalculator.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.fee.scheme.feecalculator;
+package uk.gov.justice.laa.fee.scheme.feecalculator.fixed;
 
 import static uk.gov.justice.laa.fee.scheme.feecalculator.utility.FeeCalculationUtility.calculate;
