+++ scheme-service/src/test/java/uk/gov/justice/laa/fee/scheme/feecalculator/fixed/MediationFeeCalculatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.fee.scheme.feecalculator;
+package uk.gov.justice.laa.fee.scheme.feecalculator.fixed;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
